2da parte Capa Api/Presentacion

+++ ShopPro.Web/Models/OrderDetails/OrderDetailsSaveModel.cs
@@ -4,8 +4,8 @@
     {
         public int orderid { get; set; }
         public int productid { get; set; }
-        public int unitprice { get; set; }
-        public int qty { get; set; }
-        public int discount { get; set; }
+        public decimal unitprice { get; set; }
+        public short qty { get; set; }
+        public decimal discount { get; set; }
     }
 }
